test: cover workspace avatar plan mapping through a pure util
happy-dom drops oklch() background values, so the plan color cannot be
observed through a rendered component in unit tests. Move the color
logic into workspaceAvatarStyle and test the tier mapping there: the
neutral case, the Free fallback for null and unknown tiers, and the
legacy STANDARD and FOUNDERS_EDITION aliases for Creator. The component
test keeps the wiring check that a known tier applies an inline style.
